+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Upcoming
 ---
 #### Network
+- Added rho as a network configuration
+- Added rmsDecay as a network configuration
 - Added adaptiveLR as a network configuration
 #### NetMath
 - Added adam as adaptiveLR configuration